multiple users

+++ src/main/java/org/drewandjen/dao/MasterListDaoH2.java
@@ -18,20 +18,20 @@ public class MasterListDaoH2 implements MasterListDao {
     }
 
     @Override
-    public List<MasterListItem> fetchAll() {
-        return template.query("SELECT id, name, category FROM master_list ORDER BY category", (rs, rowNum) -> {
+    public List<MasterListItem> fetchAll(Integer userId) {
+        return template.query("SELECT id, name, category FROM master_list where user_id = ? ORDER BY category", (rs, rowNum) -> {
             return new MasterListItem(rs.getInt("id"), rs.getString("name"), rs.getString("category"));
-        });
+        }, userId);
     }
 
     @Override
-    public void save(MasterListItem newItem) {
-        template.update("insert into master_list (name, category, created_at) values (?, ?, ?)",
-                newItem.getName(), newItem.getCategory(), new Date());
+    public void save(MasterListItem newItem, Integer userId) {
+        template.update("insert into master_list (name, category, user_id, created_at) values (?, ?, ?, ?)",
+                newItem.getName(), newItem.getCategory(), userId, new Date());
     }
 
     @Override
-    public void delete(MasterListItem itemToDelete) {
-        template.update("delete from master_list where id = ?", itemToDelete.getId());
+    public void delete(MasterListItem itemToDelete, Integer userId) {
+        template.update("delete from master_list where id = ? and user_id", itemToDelete.getId(), userId);
     }
 }
