+++ src/main/java/org/drewandjen/dao/ShoppingListDaoH2.java
@@ -48,27 +48,28 @@ public class ShoppingListDaoH2 implements ShoppingListDao{
     }
 
     @Override
-    public List<ShoppingList> fetchAllShoppingLists() {
-        return template.query("select id, name, created_at from shopping_list order by name", (rs, i) -> {
+    public List<ShoppingList> fetchAllShoppingLists(Integer userId) {
+        return template.query("select id, name, created_at from shopping_list where user_id = ? order by name", (rs, i) -> {
             return new ShoppingList(rs.getInt("id"), rs.getString("name"), rs.getDate("created_at"));
-        });
+        }, userId);
     }
 
     @Override
-    public void saveShopingList(String shoppingListName) {
-        template.update("insert into shopping_list(name, created_at) values(?, CURRENT_TIMESTAMP())", shoppingListName);
+    public void saveShopingList(String shoppingListName, Integer userId) {
+        template.update("insert into shopping_list(name, user_id, created_at) values(?, ?, CURRENT_TIMESTAMP())", shoppingListName, userId);
     }
 
     @Override
-    public void deleteShoppingListItem(int id) {
-        template.update("delete from shopping_list_item where id = ?", id);
+    public void deleteShoppingListItem(int id, Integer userId) {
+        template.update("delete from shopping_list_item where id = ? and user_id = ?", id, userId);
     }
 
     @Override
-    public void updateShoppingListItems(List<ShoppingListItem> itemsToUpdate) {
+    public void updateShoppingListItems(List<ShoppingListItem> itemsToUpdate, Integer userId) {
         for (ShoppingListItem shoppingListItem : itemsToUpdate) {
-            template.update("update shopping_list_item set completed = ?, name = ?, comment = ? where id = ?",
-                    shoppingListItem.isCompleted(), shoppingListItem.getName(), shoppingListItem.getComments(), shoppingListItem.getId());
+            template.update("update shopping_list_item set completed = ?, name = ?, comment = ? where id = ? and user_id = ?",
+                    shoppingListItem.isCompleted(), shoppingListItem.getName(), shoppingListItem.getComments(),
+                    shoppingListItem.getId(), userId);
         }
     }
 }
